优化过渡逻辑处理,添加css禁用功能
