+++ src/index.tsx
@@ -1,107 +1,8 @@
 import React, { cloneElement, useLayoutEffect, useMemo, useRef, useState } from 'react';
-import { isArray, isFunction, isNull, isNumber, isObject, listenerTransitionEnd, TransitionType, useLatest } from './util';
+import { isArray, isFunction, isNull, isNumber, useLatest } from './util';
 import { diff, DiffItem, StatusDefault, StatusRemove, StatusCreate } from './diff'
-
-
-const TransitionClassName = [
-    'enterClass',
-    'leaveClass',
-    'appearClass',
-    'enterToClass',
-    'leaveToClass',
-    'appearToClass',
-    'enterActiveClass',
-    'leaveActiveClass',
-    'appearActiveClass',
-] as const;
-const TransitionEventName = [
-    'onBeforeEnter',
-    'onBeforeLeave',
-    'onBeforeAppear',
-    'onEnter',
-    'onLeave',
-    'onAppear',
-    'onAfterEnter',
-    'onAfterLeave',
-    'onAfterAppear',
-    'onTransitionEnd',
-] as const;
-type TransitionTuple<T extends ReadonlyArray<string | number>, J, K = null, Y = null> = {
-    [a in T[number]]?: a extends K ? Y : J;
-};
-type Duration = number | { leave?: number, enter?: number }
-type TransitionTupleFloat<T extends ReadonlyArray<string | number>> = T[number];
-type TransitionMode = 'out-in' | 'in-out' | 'time';
-type TransitionModeTime = number;
-// eslint-disable-next-line no-redeclare
-type TransitionClassName = TransitionTuple<typeof TransitionClassName, string>;
-type TransitionEvent = TransitionTuple<typeof TransitionEventName, (el: HTMLElement) => void, 'onEnter' | 'onLeave', (el: HTMLElement, done: Function) => void>;
-type TransitionEventNameType = TransitionTupleFloat<typeof TransitionEventName>;
-interface Transition {
-    (
-        props: {
-            children?: any;
-            name?: string;
-            css?: boolean;
-            type?: TransitionType;
-            appear?: boolean;
-            mode?: TransitionMode;
-            duration?: Duration;
-            /**动画过程中的样式，加载到执行元素上 */
-            activeStyle?: React.CSSProperties | ((status: 'default' | 'enter' | 'leave') => React.CSSProperties);
-            /**
-             * 如果设置了mode默认是根据上一个组件动画结束切换，可以通过设置此属性修改组件进程离场时机
-             */
-            modeTime?: TransitionModeTime;
-            /**
-             * 设置absolute之后，会在元素上层创建一个relative属性元素
-             */
-            absolute?: boolean;
-            /**绝对定位情况下样式，加载到执行元素上一次创建带有relative的元素上 */
-            absoluteStyle?: React.CSSProperties;
-        } & TransitionClassName &
-            TransitionEvent,
-    ): JSX.Element;
-    Children: (
-        props: {
-            children?: JSX.Element | never[] | ((status: 'default' | 'enter' | 'leave', active: boolean) => JSX.Element);
-            status?: 'remove' | 'new' | 'old';
-            css?: boolean;
-            name?: string;
-            type?: TransitionType;
-            disabled?: boolean;
-            /**动画过程中的样式 */
-            activeStyle?: React.CSSProperties | ((status: 'default' | 'enter' | 'leave') => React.CSSProperties);
-            duration?: Duration;
-        } & TransitionClassName &
-            TransitionEvent,
-    ) => JSX.Element;
-}
-interface TransitionGroup {
-    (props: { children?: any; name?: string; type?: TransitionType; css?: boolean; }): JSX.Element;
-    /**
-     * 用于拓展transition-group进场离场动画效果
-     *
-     * 配合Children标签使用
-     * ```tsx
-     * <Children render={(status, active) => {
-     *    return <Css status={status} active={active}>内容</Css>
-     * }} />
-     * ```
-     */
-    Css: (props: {
-        /**正常状态样式 */
-        activeStyle?: React.CSSProperties;
-        /**进场时样式 */
-        initStyle?: React.CSSProperties;
-        /**进场时样式 */
-        style?: React.CSSProperties;
-        children?: any;
-        status: 'default' | 'enter' | 'leave';
-        active: boolean;
-        transition?: string | ((type: Parameters<TransitionGroup['Css']>[0]['status']) => string);
-    }) => JSX.Element;
-}
+import { useClassNaames, useClassNameHook, useModeEvent, useGetTransitionClass, useGetTransitionEvent, useDuration } from './hook';
+import { Transition as TransitionInterface, TransitionGroup as TransitionGroupInterface, TransitionEventNameType } from './util';
 
 const hasChangeChildrenItem = (newchild: any, oldchild: any) => {
     if (newchild.type === oldchild?.type) {
@@ -124,7 +25,7 @@ const hasChangeChildren = (newchild: any[], oldchild: any[]) => {
     return false;
 };
 
-const Transition: Transition = function (props: Parameters<Transition>[0]) {
+const Transition: TransitionInterface = function (props: Parameters<TransitionInterface>[0]) {
     const catchPrevChildren = useRef(false as any);
     const [currentChldren, setCurrentChldren] = useState(false);
     const [prevChldren, setPrevChldren] = useState(false);
@@ -187,41 +88,9 @@ const Transition: Transition = function (props: Parameters<Transition>[0]) {
         }
     }
 
-    const getTransitionClass = useMemo(() => {
-        return TransitionClassName.reduce((data: any, item: TransitionTupleFloat<typeof TransitionClassName>) => {
-            if (item === 'enterActiveClass' && init.current && props.appearActiveClass) {
-                data['enterActiveClass'] = props.appearActiveClass;
-            } else if (item === 'enterToClass' && init.current && props.appearToClass) {
-                data['enterToClass'] = props.appearToClass;
-            } else {
-                data[item] = (props as any)[item];
-            }
-            return data;
-        }, {});
-    }, [
-        props.enterToClass,
-        props.leaveToClass,
-        props.appearToClass,
-        props.enterActiveClass,
-        props.leaveActiveClass,
-        props.appearActiveClass,
-        init.current
-    ]);
+    const getTransitionClass = useGetTransitionClass(props, init);
 
-    const getTransitionEvent = useMemo(() => {
-        return TransitionEventName.reduce((data: any, item: TransitionEventNameType) => {
-            if (item === 'onBeforeEnter' && init.current && props.onBeforeAppear) {
-                data['onBeforeEnter'] = props.onBeforeAppear;
-            } else if (item === 'onAfterEnter' && init.current && props.onAfterAppear) {
-                data['onAfterEnter'] = props.onAfterAppear;
-            } else if (item === 'onEnter' && init.current && props.onAppear) {
-                data['onEnter'] = props.onAppear;
-            } else {
-                data[item] = (props as any)[item];
-            }
-            return data;
-        }, {});
-    }, [props]);
+    const getTransitionEvent = useGetTransitionEvent(props, init);
 
     const createChildren = (children: any, status: any, transitionEnd?: Function) => {
         return children ? <Children
@@ -289,7 +158,7 @@ const Transition: Transition = function (props: Parameters<Transition>[0]) {
     </div>
 } as any
 
-const Css = (props: Parameters<TransitionGroup['Css']>[0]) => {
+const Css = (props: Parameters<TransitionGroupInterface['Css']>[0]) => {
     const {
         active,
         status,
@@ -342,8 +211,8 @@ const Css = (props: Parameters<TransitionGroup['Css']>[0]) => {
     );
 };
 
-const Children: Transition['Children'] = function (
-    props: Parameters<Transition['Children']>[0],
+const Children: TransitionInterface['Children'] = function (
+    props: Parameters<TransitionInterface['Children']>[0],
 ) {
     const type = useRef<string>('old');
     const [run, setRun] = useState(false);
@@ -353,54 +222,17 @@ const Children: Transition['Children'] = function (
     const nameTransition = props.name || props.name;
     const isAnimationEnd = useRef(true);
     const [children, setChildren] = useState(null as any);
-    const latestProps = useLatest<Parameters<Transition['Children']>[0]>(props)
+    const latestProps = useLatest<Parameters<TransitionInterface['Children']>[0]>(props)
 
+
+    const classNaames = useClassNaames(props, nameTransition);
+
+    const classNameHook = useClassNameHook(props, classNaames);
+    const modeEvent = useModeEvent(props, classNaames, classNameHook);
     const hook = (name: TransitionEventNameType, done?: Function) => {
-        latestProps.current[name] && (latestProps.current as any)[name]($el.current, done);
+        // latestProps.current[name] && (latestProps.current as any)[name]($el.current, done);
+        return modeEvent(name, $el.current, done);
     };
-
-
-    const classNaames = useMemo(() => {
-        return {
-            enterClass: props.enterClass || `${nameTransition}-enter`,
-            leaveClass: props.leaveClass || `${nameTransition}-leave`,
-            enterToClass: props.enterToClass || `${nameTransition}-enter-to`,
-            leaveToClass: props.leaveToClass || `${nameTransition}-leave-to`,
-            enterActiveClass: props.enterActiveClass || `${nameTransition}-enter-active`,
-            leaveActiveClass: props.leaveActiveClass || `${nameTransition}-leave-active`,
-        };
-    }, props.css === false ? [] : [
-        nameTransition,
-        props.enterToClass,
-        props.leaveToClass,
-        props.enterActiveClass,
-        props.leaveActiveClass,
-        props.enterClass,
-        props.leaveClass
-    ]);
-
-    const classNameHook = useMemo(() => {
-        if (props.css === false) {
-            return {
-                before: '',
-                active: '',
-                to: ''
-            }
-        }
-        return {
-            before: props.status === 'remove' ? classNaames.leaveClass : classNaames.enterClass,
-            active: props.status === 'remove' ? classNaames.leaveActiveClass : classNaames.enterActiveClass,
-            to: props.status === 'new' ? classNaames.enterToClass : classNaames.leaveToClass,
-        }
-    }, props.css === false ? [] : [
-        props.status,
-        classNaames.leaveActiveClass,
-        classNaames.enterActiveClass,
-        classNaames.enterToClass,
-        classNaames.leaveToClass,
-        classNaames.leaveClass,
-        classNaames.enterClass
-    ])
 
     useLayoutEffect(() => {
         let _run = run;
@@ -411,25 +243,9 @@ const Children: Transition['Children'] = function (
             isAnimationEnd.current = false;
             requestAnimationFrame(() => {
                 hook(hookLifeCycle.before);
-                if (props.css === false) {
-                    requestAnimationFrame(enterHook);
-                } else {
-                    $el.current.classList.remove(
-                        classNaames.leaveActiveClass,
-                        classNaames.enterActiveClass,
-                        classNaames.enterToClass,
-                        classNaames.leaveToClass,
-                        classNaames.enterClass,
-                        classNaames.leaveClass
-                    )
-                    $el.current.classList.add(classNameHook.active);
-                    requestAnimationFrame(() => {
-                        // $el.current.classList.remove(classNaames.enterClass, classNaames.leaveClass);
-                        enterHook();
-                    })
-                }
+                requestAnimationFrame(enterHook);
             })
-            stopTransitionEnd.current();
+            stopTransitionEnd.current && stopTransitionEnd.current();
         }
         setRun(_run);
         type.current = props.status as any;
@@ -441,20 +257,6 @@ const Children: Transition['Children'] = function (
         return isFunction(props.children) ? (props as any).children(status, active) : props.children;
     };
 
-    const duration = useMemo(() => {
-        if (!props.duration) return 0;
-        let enter = 0;
-        let leave = 0;
-        if (isObject(props.duration)) {
-            const _duration = props.duration as any;
-            enter = _duration.enter;
-            leave = _duration.leave;
-        } else {
-            enter = leave = (props.duration as any) || 0;
-        }
-        return props.status === 'new' ? enter : leave;
-    }, [props.duration, props.status])
-
     const hookLifeCycle = useMemo(() => {
         return {
             before: props.status === 'remove' ? 'onBeforeLeave' : 'onBeforeEnter' as any,
@@ -464,12 +266,6 @@ const Children: Transition['Children'] = function (
     }, [props.status])
 
     const enterHook = () => {
-        
-        (props.css !== false) && $el.current.classList.add(classNameHook.to);
-        hook(hookLifeCycle.run, () => {
-            end();
-        });
-        setChildren(renderChildren(true, true));
         const end = () => {
             if (!$el.current) return;
             hook(hookLifeCycle.after);
@@ -477,16 +273,8 @@ const Children: Transition['Children'] = function (
             setChildren(renderChildren(false, false));
             hook('onTransitionEnd');
         }
-        if (duration) {
-            const tiemout = setTimeout(end, duration) as any
-            stopTransitionEnd.current = () => {
-                clearTimeout(tiemout);
-            };
-        } else if (props.css !== false) {
-            const { stop } = listenerTransitionEnd($el.current, props.type || 'transition', end);
-            stopTransitionEnd.current = stop as any;
-        }
-
+        stopTransitionEnd.current = hook(hookLifeCycle.run, end);
+        setChildren(renderChildren(true, true));
     }
 
     const activeStyle = useMemo(() => {
@@ -536,7 +324,7 @@ const Children: Transition['Children'] = function (
 } as any;
 
 // eslint-disable-next-line no-redeclare
-const TransitionGroup: TransitionGroup = function (props: Parameters<TransitionGroup>[0]) {
+const TransitionGroup: TransitionGroupInterface = function (props: Parameters<TransitionGroupInterface>[0]) {
     const prevChildren = useRef();
     const catchPrevChildren = useRef([]);
     const children = useMemo(() => {
